+++ openwords_server/Java_Server/OpenwordsServer/src/java/com/openwords/actions/accounts/LoginUser.java
@@ -16,7 +16,7 @@ public class LoginUser extends MyAction {
 
     private static final long serialVersionUID = 1L;
     private boolean result;
-    private int userId = -1;
+    private long userId = -1;
     private String username, password, errorMessage;
 
     @Action(value = "/loginUser", results = {
@@ -56,7 +56,7 @@ public class LoginUser extends MyAction {
         return result;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
