+++ openwords_server/Java_Server/OpenwordsServer/src/java/com/openwords/database/WordForTTS.java
@@ -2,17 +2,17 @@ package com.openwords.database;
 
 public class WordForTTS {
 
-    private int wordId;
+    private long wordId;
     private String word;
 
     public WordForTTS() {
     }
 
-    public int getWordId() {
+    public long getWordId() {
         return wordId;
     }
 
-    public void setWordId(int wordId) {
+    public void setWordId(long wordId) {
         this.wordId = wordId;
     }
